+++ src/components/explorer/ExplorerIntroState.js
@@ -8,10 +8,6 @@ import { formatNum } from '../../util/formats'
 import mapCrimeToGlossaryTerm from '../../util/glossary'
 import ucrParticipationLookup from '../../util/participation'
 
-const highlight = txt =>
-  <strong>
-    {txt}
-  </strong>
 const getReportTerms = ({ nibrs, srs, hybrid }) =>
   <span>
     {hybrid && 'both '}
@@ -20,10 +16,43 @@ const getReportTerms = ({ nibrs, srs, hybrid }) =>
     {nibrs && <NibrsTerm />}
   </span>
 
+const getStateReportingHistory = ({ nibrs, hybrid, placeName }) => {
+  if (!nibrs) return false
+
+  let text
+  if (hybrid) {
+    text = (
+      <span>
+        has historically reported <SrsTerm /> data and began including{' '}
+        <NibrsTerm /> data in {nibrs['initial-year']}
+      </span>
+    )
+  } else if (!hybrid && nibrs) {
+    text = (
+      <span>
+        historically reported <SrsTerm /> data but switched to submitting{' '}
+        <NibrsTerm /> data in {nibrs['initial-year']}
+      </span>
+    )
+  }
+
+  return (
+    <span>
+      The {placeName} state UCR Program {text}.
+    </span>
+  )
+}
+
+const highlight = txt =>
+  <strong>
+    {txt}
+  </strong>
+
 const ExplorerIntroState = ({ crime, page, place, participation, until, placeName }) => {
   const isCrime = page === 'crime'
   const isArson = crime === 'arson'
   const { nibrs, srs } = ucrParticipationLookup(place)
+  const hybrid = nibrs && srs
   const untilUcr = participation.find(p => p.year === until)
   let totalAgencies = 0
   let coveredAgencies = 0
@@ -31,15 +60,26 @@ const ExplorerIntroState = ({ crime, page, place, participation, until, placeNam
     totalAgencies = untilUcr.total_agencies
     coveredAgencies = untilUcr.covered_agencies
   }
-  console.log('ExplorerIntroState - page:', page)
 
+  let crimeTerm
+  let reportTerms
   if (isCrime) {
-    const reportTerms = getReportTerms({ nibrs, srs, hybrid: nibrs && srs })
-    const crimeTerm = (
-      <Term id={mapCrimeToGlossaryTerm(crime)}>
-        {upperFirst(lowerCase(crime))}
-      </Term>
-    )
+       reportTerms = getReportTerms({ nibrs, srs, hybrid: nibrs && srs })
+       crimeTerm = (
+        <Term id={mapCrimeToGlossaryTerm(crime)}>
+          {upperFirst(lowerCase(crime))}
+        </Term>
+      )
+
+    const stateProgram = {
+      nibrs,
+      place,
+      srs,
+      hybrid,
+    }
+
+    const programHistory = getStateReportingHistory(stateProgram, placeName)
+
 
     return (
       <div>
@@ -47,16 +87,16 @@ const ExplorerIntroState = ({ crime, page, place, participation, until, placeNam
           ? <div>
               <p className="serif">
                 {crimeTerm} rates for {placeName} are derived from{' '}
-                {reportTerms} reports voluntarily submitted to the FBI.
+                {reportTerms} reports voluntarily submitted to the FBI.{' '}
+                {programHistory}
               </p>
               <p className="serif">
                 In {highlight(until)}
-                , the FBI <EstimatedTerm /> crime statistics for{' '}
-
-                {placeName} based on data received from{' '}
-                {highlight(formatNum(totalAgencies - coveredAgencies))} law
+                , the FBI <EstimatedTerm /> crime statistics for {placeName}{' '}
+                based on data received from{' '}
+                {highlight(formatNum(untilUcr.participating_agencies))} law
                 enforcement agencies out of{' '}
-                {highlight(formatNum(totalAgencies))} agencies in the
+                {highlight(formatNum(untilUcr.total_agencies))} agencies in the
                 state that year.
               </p>
             </div>
@@ -65,30 +105,29 @@ const ExplorerIntroState = ({ crime, page, place, participation, until, placeNam
                 {placeName} reports {reportTerms} data to the FBI.
               </p>
               <p className="serif">
-
-                In {until}, {formatNum(totalAgencies - coveredAgencies)}{' '}
-                {placeName} law enforcement agencies voluntarily
-                reported data to the FBI. The charts below feature unestimated
-                data.
+                In {until}, {formatNum(untilUcr.participating_agencies)}{' '}
+                {placeName} law enforcement agencies voluntarily reported data
+                to the FBI. The charts below feature unestimated data.
               </p>
             </div>}
       </div>
     )
   }
-  return (
-    <div>
-      <p className="serif">
-        In {highlight(until)}
-        , the FBI <EstimatedTerm /> crime statistics for{' '}
 
-        {placeName} based on data received from{' '}
-        {highlight(formatNum(totalAgencies - coveredAgencies))} law
-        enforcement agencies out of{' '}
-        {highlight(formatNum(totalAgencies))} agencies in the
-        state that year.
-      </p>
-    </div>
-  )
+  return (
+   <div>
+     <p className="serif">
+       In {highlight(until)}
+       , the FBI <EstimatedTerm /> crime statistics for{' '}
+
+       {placeName} based on data received from{' '}
+       {highlight(formatNum(totalAgencies - coveredAgencies))} law
+       enforcement agencies out of{' '}
+       {highlight(formatNum(totalAgencies))} agencies in the
+       state that year.
+     </p>
+   </div>
+ )
 }
 
 ExplorerIntroState.propTypes = {}