+++ src/server.js
@@ -3,16 +3,20 @@
 import 'babel-polyfill'
 import 'newrelic'
 
+import path from 'path'
+import url from 'url'
+
 import http from 'axios'
 import bodyParser from 'body-parser'
 import express from 'express'
 import gzipStatic from 'connect-gzip-static'
-import path from 'path'
+import helmet from 'helmet'
 import React from 'react'
-import Helmet from 'react-helmet'
+import ReactHelmet from 'react-helmet'
 import { renderToString } from 'react-dom/server'
 import { Provider } from 'react-redux'
 import { match, RouterContext } from 'react-router'
+import uuid from 'uuid/v4'
 
 import packageJson from '../package.json'
 import renderHtml from './html'
@@ -24,7 +28,6 @@ import { fetchUcrRegion } from './actions/region'
 import { fetchUcrState } from './actions/states'
 
 import createEnv from './util/env'
-import { hasThreatKeyword, notifyOfThreat } from './util/feedback'
 import { createIssue } from './util/github'
 import history from './util/history'
 
@@ -39,7 +42,6 @@ const {
   GITHUB_ISSUE_REPO_NAME: repoName,
   GITHUB_ISSUE_BOT_TOKEN: repoToken,
   PORT,
-  THREAT_KEYWORDS,
 } = ENV
 
 const initState = {
@@ -58,21 +60,46 @@ const acceptHostname = hostname => {
 const app = express()
 
 const publicDirPath = path.join(__dirname, '..', 'public')
+app.use((req, res, next) => {
+  /* eslint-disable no-param-reassign */
+  res.locals.nonce = uuid()
+  /* eslint-enable no-param-reassign */
+  next()
+})
 app.use(gzipStatic(__dirname))
 app.use(gzipStatic(publicDirPath))
 app.use(bodyParser.json())
 
+const defaultSrc = [
+  "'self'", // serve any assets from this server
+  url.parse(API).hostname, // enable any requests to the API server
+  'www.google-analytics.com',
+  'dap.digitalgov.gov',
+]
+app.use(helmet())
+app.use(
+  helmet.contentSecurityPolicy({
+    directives: {
+      defaultSrc,
+      scriptSrc: [
+        ...defaultSrc,
+        "'unsafe-eval'", // unfortunately, required for DAP right now
+        (req, res) => `'nonce-${res.locals.nonce}'`,
+      ],
+      styleSrc: [...defaultSrc, "'unsafe-inline'"],
+    },
+  }),
+)
+
 app.get('/api', (req, res) => {
   res.sendfile('/swagger/index.html', { root: publicDirPath })
 })
-
 
 app.get('/api-proxy/*', (req, res) => {
   const route = `${API}/${req.params['0']}`.replace(/\/$/g, '')
   const params = Object.assign({}, req.query, { api_key: apiKey })
 
   if (!apiKey) return res.status(401).end()
-
 
   return http
     .get(route, { params })
@@ -87,7 +114,6 @@ app.get('/api-proxy/*', (req, res) => {
 
 app.post('/feedback', (req, res) => {
   const { body, title } = req.body
-  const terms = THREAT_KEYWORDS && JSON.parse(THREAT_KEYWORDS)
   const allEnvs = repoOwner && repoName && repoToken
 
   if (!allEnvs || !acceptHostname(req.hostname)) return res.status(401).end()
@@ -99,11 +125,7 @@ app.post('/feedback', (req, res) => {
     title,
     token: repoToken,
   })
-    .then(issueResponse => {
-      const { data: issue } = issueResponse
-      if (hasThreatKeyword(body, terms)) notifyOfThreat(issue)
-      return res.send(issue)
-    })
+    .then(issue => res.send(issue.data))
     .catch(e => res.status(e.response.status).end())
 })
 
@@ -129,9 +151,8 @@ app.get('/*', (req, res) => {
           <RouterContext {...props} />
         </Provider>,
       )
-      const head = Helmet.rewind()
-
-      res.send(renderHtml(html, head, store.getState()))
+      const head = ReactHelmet.rewind()
+      res.send(renderHtml(html, head, store.getState(), res.locals.nonce))
     }
   })
 })