+++ src/util/participation.js
@@ -1,6 +1,6 @@
 import offensesUtil from './offenses'
 import { oriToState } from './agencies'
-import { validateFilter, isValidState } from './location'
+import { validateFilter } from './location'
 
 import data from '../../public/data/ucr-program-participation.json'
 
@@ -14,7 +14,7 @@ export const shouldFetchUcr = (filters, region, states) => filters.placeType !==
 export const shouldFetchSummaries = (filters, region, states) =>
   isValidCrime(filters.pageType) && validateFilter(filters, region.regions, states.states)
 
-export const shouldFetchNibrs = ({ crime, place, placeType }, states) => {
+export const shouldFetchNibrs = ({ crime, place, placeType }) => {
   if (noNibrs.includes(crime) || placeType === 'region') return false
 
   const placeNorm = placeType === 'agency' ? oriToState(place) : place